Move "reporting beta" page to "reporting"
And remove the now redundant code for calling metrics endpoints in workflow.
