--- pages/reporting-beta/content/index.js
@@ -1,30 +1,0 @@
-module.exports = {
-  breadcrumbs: {
-    reportingBeta: 'Performance metrics'
-  },
-  'legacy-project-application': 'PPL applications (Legacy)',
-  'legacy-project-amendment': 'PPL amendments (Legacy)',
-  'legacy-project-revoke': 'PPL revocations (Legacy)',
-  'legacy-project-transfer': 'PPL transfers (Legacy)',
-  'project-application': 'PPL applications (New)',
-  'project-amendment': 'PPL amendments (New)',
-  'project-revoke': 'PPL revocations (New)',
-  'project-transfer': 'PPL transfers (New)',
-  'pil-application': 'PIL applications',
-  'pil-amendment': 'PIL amendments',
-  'pil-revoke': 'PIL revocations',
-  'pil-transfer': 'PIL transfers',
-  'pil-review': 'PIL reviews',
-  'role-create': 'Named people assignments',
-  'role-delete': 'Named people revocations',
-  'place-update': 'Approved area amendments',
-  'place-create': 'Approved area additions',
-  'place-delete': 'Approved area deletions',
-  'profile-update': 'Profile updates',
-  date: {
-    all: 'All',
-    week: 'This week',
-    month: 'This month',
-    year: 'This year'
-  }
-};
