--- pages/reporting-beta/routes.js
@@ -1,9 +1,0 @@
-const download = require('./download');
-
-module.exports = {
-  download: {
-    breadcrumb: false,
-    path: '/download',
-    router: download
-  }
-};
