+++ pages/reporting/content/index.js
@@ -1,8 +1,12 @@
 module.exports = {
-  'project-application': 'PPL applications',
-  'project-amendment': 'PPL amendments',
-  'project-revoke': 'PPL revocations',
-  'project-transfer': 'PPL transfers',
+  'legacy-project-application': 'PPL applications (Legacy)',
+  'legacy-project-amendment': 'PPL amendments (Legacy)',
+  'legacy-project-revoke': 'PPL revocations (Legacy)',
+  'legacy-project-transfer': 'PPL transfers (Legacy)',
+  'project-application': 'PPL applications (New)',
+  'project-amendment': 'PPL amendments (New)',
+  'project-revoke': 'PPL revocations (New)',
+  'project-transfer': 'PPL transfers (New)',
   'pil-application': 'PIL applications',
   'pil-amendment': 'PIL amendments',
   'pil-revoke': 'PIL revocations',
