+++ pages/reporting/index.js
@@ -1,7 +1,16 @@
 const { page } = require('@asl/service/ui');
-const downloads = require('./downloads');
+const { pipeline } = require('stream');
+const through = require('through2');
+
+const metrics = require('../../lib/middleware/metrics');
+
+const routes = require('./routes');
 
 const types = [
+  'legacy-project-application',
+  'legacy-project-amendment',
+  'legacy-project-revoke',
+  'legacy-project-transfer',
   'project-application',
   'project-amendment',
   'project-revoke',
@@ -25,22 +34,65 @@ module.exports = settings => {
     root: __dirname
   });
 
+  app.use(metrics(settings));
+
+  app.use('/', (req, res, next) => {
+    res.locals.static.since = req.query.since || '2019-07-01';
+    res.locals.static.types = types;
+    next();
+  });
+
   app.get('/', (req, res, next) => {
-    const since = req.query.since || '2019-07-01';
-    req.api(`/metrics?since=${since}`)
-      .then(response => {
-        res.locals.static.since = since;
-        res.locals.static.metrics = response.json;
-        res.locals.static.types = types;
+    req.metrics('/active-licences', { stream: false })
+      .then(json => {
+        res.locals.static.licences = json;
         next();
       })
       .catch(next);
-
   });
 
-  app.use(downloads({ types }));
+  app.get('/ppl-sla', (req, res, next) => {
+    req.metrics('/reports/ppl-sla', { stream: false, query: req.query })
+      .then(response => {
+        res.json(response);
+      })
+      .catch(next);
+  });
+
+  app.get('/tasks', (req, res, next) => {
+    req.metrics('/reports/tasks', { stream: true, query: req.query })
+      .then(stream => {
+        const tasks = {
+          total: 0
+        };
+        pipeline(
+          stream,
+          through.obj((data, enc, callback) => {
+            tasks.total++;
+            let type = `${data.model}-${data.action}`;
+            if (data.model === 'project' && data.schemaVersion === 0) {
+              type = `legacy-project-${data.action}`;
+            }
+            tasks[type] = tasks[type] + 1 || 1;
+            if (data.iterations) {
+              tasks[`${type}-iterations`] = tasks[`${type}-iterations`] + data.iterations || data.iterations;
+            }
+            callback();
+          }),
+          err => {
+            if (err) {
+              return next(err);
+            }
+            res.json(tasks);
+          }
+        );
+      })
+      .catch(next);
+  });
 
   app.get('/', (req, res) => res.sendResponse());
 
   return app;
 };
+
+module.exports.routes = routes;
